+++ src/video/texture/cache.js
@@ -88,7 +88,7 @@ class TextureCache {
 
         if (typeof entry === "undefined") {
             if (!atlas) {
-                atlas = createAtlas(image.width, image.height, image.src ? fileUtil.getBasename(image.src) : undefined);
+                atlas = createAtlas(image.width || image.videoWidth, image.videoHeight, image.src ? fileUtil.getBasename(image.src) : undefined);
             }
             entry = new TextureAtlas(atlas, image, false);
             this.set(image, entry);
@@ -132,8 +132,8 @@ class TextureCache {
      * @ignore
      */
     set(image, texture) {
-        let width = image.width;
-        let height = image.height;
+        let width = image.width || image.videoWidth;
+        let height = image.height || image.videoHeight;
 
         // warn if a non POT texture is added to the cache when using WebGL1
         if (renderer.WebGLVersion === 1 && (!isPowerOfTwo(width) || !isPowerOfTwo(height))) {